# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VehiGo API docs",
        "description": "This is an API documentation, generated by L5 Swagger",
        "contact": {
            "email": "office@atelierulmeu.ro"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/cars/{id}": {
            "get": {
                "tags": [
                    "Car"
                ],
                "summary": "Get user car by id",
                "description": "Returns the car detailed data",
                "operationId": "e8cc4357610811cddead04829ec723f9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of car",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Car response by car id.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CarResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_Auth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Car"
                ],
                "summary": "Update the current car",
                "description": "Updates the user car by id",
                "operationId": "81779f5cda2fd8b16e677b7584ae243e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of car",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Car fillable properties",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated car response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CarResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_Auth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Car"
                ],
                "summary": "Delete the user car",
                "description": "Deletes the user car by id",
                "operationId": "2358187c3e9f97486628f50543ec541a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of car",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete car response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteCarResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_Auth": []
                    }
                ]
            }
        },
        "/cars": {
            "get": {
                "tags": [
                    "Car"
                ],
                "summary": "Get user cars",
                "description": "Returns the user related cars",
                "operationId": "4abf8c869538cb6e4644537ba31a6aad",
                "responses": {
                    "200": {
                        "description": "The user cars response.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserCarsResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_Auth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Car"
                ],
                "summary": "Create new user car",
                "description": "Creates a new user related car",
                "operationId": "eefaab256ebfc761559e49bda6467817",
                "requestBody": {
                    "description": "Car fillable properties",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Car response by car id.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CarResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_Auth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CarBody": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "My Car"
                    },
                    "category": {
                        "type": "string",
                        "example": "Autoturism/Automobil mixt"
                    },
                    "subcategory": {
                        "type": "string",
                        "example": "Autoturism"
                    },
                    "registration_type": {
                        "type": "string",
                        "example": "Inmatriculat"
                    },
                    "fuel_type": {
                        "type": "string",
                        "example": "Motorina"
                    },
                    "vin_number": {
                        "type": "string",
                        "example": "WBAAP71111GL33030"
                    },
                    "make": {
                        "type": "string",
                        "example": "BMW"
                    },
                    "model": {
                        "type": "string",
                        "example": "Seria 3"
                    },
                    "manufacture_year": {
                        "type": "string",
                        "example": "2002"
                    },
                    "tyre_size": {
                        "type": "json",
                        "example": []
                    },
                    "motor_power": {
                        "type": "string",
                        "example": "110"
                    },
                    "cylinder_capacity": {
                        "type": "string",
                        "example": "1995"
                    },
                    "number_places": {
                        "type": "string",
                        "example": "5"
                    },
                    "max_per_mass": {
                        "type": "string",
                        "example": "1550"
                    },
                    "civ_number": {
                        "type": "string",
                        "example": "K884163"
                    },
                    "description": {
                        "type": "string",
                        "example": "My favorite car"
                    },
                    "favorite": {
                        "type": "boolean",
                        "example": "true"
                    }
                },
                "type": "object"
            },
            "CarResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": ""
                    },
                    "data": {
                        "$ref": "#/components/schemas/Car"
                    }
                },
                "type": "object"
            },
            "DeleteCarResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": ""
                    },
                    "data": {
                        "type": "object",
                        "example": []
                    }
                },
                "type": "object"
            },
            "UserCarsResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": ""
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Car"
                        }
                    }
                },
                "type": "object"
            },
...
            "Car": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "My Car"
                    },
                    "category": {
                        "type": "string",
                        "example": "Autoturism/Automobil mixt"
                    },
                    "subcategory": {
                        "type": "string",
                        "example": "Autoturism"
                    },
                    "registration_type": {
                        "type": "string",
                        "example": "Inmatriculat"
                    },
                    "fuel_type": {
                        "type": "string",
                        "example": "Motorina"
                    },
                    "vin_number": {
                        "type": "string",
                        "example": "WBAAP71111GL33030"
                    },
                    "make": {
                        "type": "string",
                        "example": "BMW"
                    },
                    "model": {
                        "type": "string",
                        "example": "Seria 3"
                    },
                    "manufacture_year": {
                        "type": "string",
                        "example": "2002"
                    },
                    "tyre_size": {
                        "type": "json",
                        "example": []
                    },
                    "motor_power": {
                        "type": "string",
                        "example": "110"
                    },
                    "cylinder_capacity": {
                        "type": "string",
                        "example": "1995"
                    },
                    "number_places": {
                        "type": "string",
                        "example": "5"
                    },
                    "max_per_mass": {
                        "type": "string",
                        "example": "1550"
                    },
                    "civ_number": {
                        "type": "string",
                        "example": "K884163"
                    },
                    "description": {
                        "type": "string",
                        "example": "My favorite car"
                    },
                    "favorite": {
                        "type": "boolean",
                        "example": "true"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer_Auth": {
                "type": "apiKey",
                "description": "A short description for security scheme",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}